+++ packages/core/components/auth/auth-listener.tsx
@@ -6,6 +6,11 @@ import { useAuth } from '@hanzo/auth/service'
 const AuthListener = () => {
   const auth = useAuth()
 
+  const requestAuthToken = () => {
+    const childFrame = document.getElementById('login') as HTMLIFrameElement
+    childFrame?.contentWindow?.postMessage(true, process.env.NEXT_PUBLIC_LOGIN_SITE_URL ?? '')
+  }
+
   useEffect(() => {
     if (!!localStorage.getItem('auth-token')) {
       console.log("Detected auth-token cookie... Logging in...")
@@ -16,6 +21,7 @@ const AuthListener = () => {
       if (event.origin === process.env.NEXT_PUBLIC_AUTH_ORIGIN) {
         const token = event.data
         localStorage.setItem('auth-token', token)
+        auth.loginWithCustomToken(localStorage.getItem('auth-token') as string)
       }
     }
 
@@ -26,7 +32,7 @@ const AuthListener = () => {
     }
   }, [])
 
-  return (<></>)
+  return (<iframe id='login' onLoad={requestAuthToken} src={process.env.NEXT_PUBLIC_LOGIN_SITE_URL}/>)
 }
 
 export default AuthListener
