+++ sites/login/src/app/layout.tsx
@@ -3,16 +3,11 @@ import React, { type PropsWithChildren } from 'react'
 import { 
   RootLayout as RootLayoutCommon, 
   rootLayoutViewport, 
-  ChatWidget
 } from '@luxfi/core'
-
-import { Toaster } from '@hanzo/ui/primitives'
 
 import { AuthServiceProvider } from '@hanzo/auth/service'
 import { getUserServerSide } from '@hanzo/auth/server'
 import type { AuthServiceConf } from '@hanzo/auth/types'
-
-import { CommerceServiceProvider } from '@hanzo/commerce'
 
 import siteDef from '../site-def'
 import _metadata from '../metadata'
@@ -31,11 +26,6 @@ const RootLayout: React.FC<PropsWithChildren> = async ({
     <AuthServiceProvider user={currentUser} conf={{} as AuthServiceConf}>
       <RootLayoutCommon siteDef={siteDef} header={false} >
         {children}
-        <ChatWidget
-          title='LUX'
-          subtitle='AI'
-          chatbotUrl='https://lux.chat/iframe'
-        />
       </RootLayoutCommon>
     </AuthServiceProvider>
   )
