+++ sites/login/src/app/page.tsx
@@ -1,48 +1,81 @@
 'use client'
 
-import React  from 'react'
-import { observer } from 'mobx-react-lite'
-
-import { ApplyTypography, Main } from '@hanzo/ui/primitives'
-import { useAuth } from '@hanzo/auth/service'
-import { LoginPanel } from '@hanzo/auth/components'
-
-import { AccessCodeInput, Footer, Header } from '@luxfi/core'
-
-import siteDef from '../site-def'
 import { useRouter } from 'next/navigation'
 
-const Home = observer(() => {
-  const router = useRouter()
-  const auth = useAuth()
+import { ChatWidget, LoginPanel } from '@luxfi/core'
 
-  const validCodes = ['777777']
-
-  const redirectToApp = () => {
-    if (typeof window !== 'undefined') {
-      const { protocol, hostname } = window.location
-      const newHostname = `app.${hostname}`
-      router.push(`${protocol}//${newHostname}`)
-    }
+const reviews = [
+  {
+    text: 'Lux Defi is amazing!',
+    author: 'Giovanna Mingarelli',
+    href: 'https://trstp.lt/a55NNi_j9'
+  },
+  {
+    text: 'Lux is revolutionizing the blockchain world by seamlessly merging commodities like gold and silver with digital technology.',
+    author: 'Giovanna Mingarelli',
+    href: 'https://trstp.lt/a55NNi_j9'
+  },
+  {
+    text: 'Lux Credit Card has the ability to make transactions that let you borrow from any of your assets in real time! A total game-changer and I highly recommend it.',
+    author: 'Giovanna Mingarelli',
+    href: 'https://trstp.lt/a55NNi_j9'
+  },
+  {
+    text: 'Lux is the best! Highly recommend this great service for anyone interested in DeFi.',
+    author: 'Ben Chu',
+    href: 'https://trstp.lt/PtpL8qO-U'
+  },
+  {
+    text: 'Lux is better than several competitors, good interface. Big fan of the product team and the mission.',
+    author: 'Ben Chu',
+    href: 'https://trstp.lt/PtpL8qO-U'
+  },
+  {
+    text: 'Setting the bar for innovation, development and execution within an ever evolving mix of AI, Blockchain, Design and Finance.',
+    author: 'Ole Brereton',
+    href: 'https://trstp.lt/gOB3GTbOb'
+  },
+  {
+    text: 'World class leaders in the future of metaverse and integrated, financial systems.',
+    author: 'Ole Brereton',
+    href: 'https://trstp.lt/gOB3GTbOb'
+  },
+  {
+    text: 'As an art collector and real estate executive, I value unique investments. Lux Defi, merging real-world assets with blockchain, is a refreshing addition.',
+    author: 'Lisa Goodman',
+    href: 'https://trstp.lt/GfjNiLa7D'
+  },
+  {
+    text: 'Lux, merging real-world assets with blockchain, is a refreshing addition alongside my art works and other assets.',
+    author: 'Lisa Goodman',
+    href: 'https://trstp.lt/GfjNiLa7D'
   }
+]
 
+type Props = {
+  searchParams?: { [key: string]: string | string[] | undefined }
+}
+
+const Page = ({ searchParams }: Props ) => {
+  const router = useRouter()
+
+  const handleLoginDone = () => {
+    router.back()
+  }
+  // max-w-screen-2xl 2xl:w-[1500px]
   return (<>
-    <Header siteDef={siteDef} />
-    <Main className='h-screen -mt-[44px] md:-mt-[80px]'>
-      <ApplyTypography className='flex flex-col gap-8 m-auto text-center'>
-        <h1>MEMBER LOGIN</h1>
-        {auth.loggedIn ? (
-          <div className='flex flex-col gap-4 max-w-[40rem] mx-auto'>
-            <h5>If you are an existing Lux member accessing this website for the first time, please enter your invite code.</h5>
-            <AccessCodeInput validCodes={validCodes} onSuccess={redirectToApp}/>
-          </div>
-        ) : (
-          <LoginPanel noHeading className='max-w-[20rem] mx-auto' />
-        )}
-      </ApplyTypography>
-    </Main>
-    <Footer siteDef={siteDef} className='max-w-screen-2xl w-full pt-16 lg:mx-auto ' />
+    <LoginPanel
+      close={handleLoginDone}
+      getStartedUrl='/'
+      redirectUrl={searchParams?.redirectUrl as string ?? undefined}
+      reviews={reviews}
+    />
+    <ChatWidget
+      title='LUX'
+      subtitle='AI'
+      chatbotUrl='https://lux.chat/iframe'
+    />
   </>)
-})
+}
 
-export default Home
+export default Page
