--- sites/market/src/app/login/page.tsx
@@ -1,2 +1,0 @@
-const Page = () => ( <></> )
-export default Page
